# xiaoyang-web-http-pb-152026/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "xiaoyang api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/leagues": {
      "post": {
        "summary": "创建组织",
        "description": "创建组织",
        "operationId": "LeaguesService_CreateLeague",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateLeagueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateLeagueRequest"
            }
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": [],
            "LeagueAuth": []
          }
        ]
      }
    },
    "/v1/leagues/list": {
      "get": {
        "summary": "获取组织列表",
        "description": "获取组织列表",
        "operationId": "LeaguesService_ListLeagues",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListLeaguesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "description": "用户ID",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "qqNumber",
            "description": "用户QQ号",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/leagues/{ukey}": {
      "get": {
        "summary": "获取组织详情",
        "description": "获取组织详情",
        "operationId": "LeaguesService_GetLeague",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetLeagueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "description": "组织ID",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新组织",
        "description": "更新组织",
        "operationId": "LeaguesService_UpdateLeague",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateLeagueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LeaguesServiceUpdateLeagueBody"
            }
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/leagues/{ukey}/expired_at": {
      "put": {
        "summary": "更新组织过期时间",
        "description": "更新组织过期时间",
        "operationId": "LeaguesService_UpdateLeagueExpiredAt",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateLeagueExpiredAtResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LeaguesServiceUpdateLeagueExpiredAtBody"
            }
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/leagues/{ukey}/join": {
      "post": {
        "summary": "加入组织",
        "description": "加入组织",
        "operationId": "LeaguesService_JoinLeague",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1JoinLeagueResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LeaguesServiceJoinLeagueBody"
            }
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/leagues/{ukey}/members": {
      "get": {
        "summary": "获取组织成员列表",
        "description": "获取组织成员列表",
        "operationId": "LeaguesService_ListLeagueMembers",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListLeagueMembersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "groupId",
            "description": "组织ID",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/leagues/{ukey}/user_nickname": {
      "put": {
        "summary": "修改用户在组织中的昵称",
        "description": "修改用户在组织中的昵称",
        "operationId": "LeaguesService_SetUserNickname",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetUserNicknameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LeaguesServiceSetUserNicknameBody"
            }
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/leagues/{ukey}/user_role": {
      "put": {
        "summary": "设定用户在组织中的角色",
        "description": "设定用户在组织中的角色",
        "operationId": "LeaguesService_SetUserRole",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetUserRoleResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ukey",
            "description": "群唯一标识",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LeaguesServiceSetUserRoleBody"
            }
          }
        ],
        "tags": [
          "LeaguesService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "LeaguesServiceJoinLeagueBody": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64",
          "title": "用户ID"
        },
        "qqNumber": {
          "type": "string",
          "title": "用户QQ号"
        },
        "groupNumber": {
          "type": "string",
          "title": "群号"
        },
        "groupId": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "role": {
          "type": "string",
          "title": "角色"
        },
        "groupNickname": {
          "type": "string",
          "title": "群昵称"
        }
      },
      "title": "加入组织请求"
    },
    "LeaguesServiceSetUserNicknameBody": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64",
          "title": "用户ID"
        },
        "qqNumber": {
          "type": "string",
          "title": "用户QQ号"
        },
        "groupId": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "nickname": {
          "type": "string",
          "title": "昵称"
        }
      },
      "title": "修改用户在组织中的昵称请求"
    },
    "LeaguesServiceSetUserRoleBody": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64",
          "title": "用户ID"
        },
        "qqNumber": {
          "type": "string",
          "title": "用户QQ号"
        },
        "groupId": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "role": {
          "type": "string",
          "title": "角色"
        }
      },
      "title": "设定用户在组织中的角色请求"
    },
    "LeaguesServiceUpdateLeagueBody": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "groupNumber": {
          "type": "string",
          "title": "群号"
        },
        "name": {
          "type": "string",
          "title": "群名称"
        },
        "server": {
          "type": "string",
          "title": "服务器"
        },
        "avatar": {
          "type": "string",
          "title": "群头像"
        },
        "preferences": {
          "type": "string",
          "title": "偏好设置"
        }
      },
      "title": "更新组织请求"
    },
    "LeaguesServiceUpdateLeagueExpiredAtBody": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "expiredAt": {
          "type": "integer",
          "format": "int64",
          "title": "过期时间"
        }
      },
      "title": "更新组织过期时间请求"
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1CreateLeagueRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64",
          "title": "创建人ID(admin)"
        },
        "qqNumber": {
          "type": "string",
          "title": "创建人QQ号(admin)"
        },
        "groupNumber": {
          "type": "string",
          "title": "群号"
        },
        "ukey": {
          "type": "string",
          "title": "群唯一标识"
        },
        "name": {
          "type": "string",
          "title": "群名称"
        },
        "server": {
          "type": "string",
          "title": "服务器"
        },
        "avatar": {
          "type": "string",
          "title": "群头像"
        },
        "expiredAt": {
          "type": "integer",
          "format": "int64",
          "title": "过期时间"
        },
        "preferences": {
          "type": "string",
          "title": "偏好设置"
        }
      },
      "title": "创建组织请求"
    },
    "v1CreateLeagueResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        }
      },
      "title": "创建组织响应"
    },
...
    "v1GetLeagueResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "groupNumber": {
          "type": "string",
          "title": "群号"
        },
        "ukey": {
          "type": "string",
          "title": "群唯一标识"
        },
        "name": {
          "type": "string",
          "title": "群名称"
        },
        "server": {
          "type": "string",
          "title": "服务器"
        },
        "avatar": {
          "type": "string",
          "title": "群头像"
        },
        "expiredAt": {
          "type": "integer",
          "format": "int64",
          "title": "过期时间"
        },
        "preferences": {
          "type": "string",
          "title": "偏好设置"
        }
      },
      "title": "获取组织详情响应"
    },
...
    "v1JoinLeagueResponse": {
      "type": "object",
      "title": "加入组织响应"
    },
    "v1LeagueMemberSummary": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64",
          "title": "用户ID"
        },
        "qqNumber": {
          "type": "string",
          "title": "用户QQ号"
        },
        "nickname": {
          "type": "string",
          "title": "昵称"
        },
        "avatar": {
          "type": "string",
          "title": "头像"
        },
        "role": {
          "type": "string",
          "title": "角色"
        },
        "groupNickname": {
          "type": "string",
          "title": "群昵称"
        }
      },
      "title": "组织成员信息摘要"
    },
    "v1LeagueSummary": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "title": "组织ID"
        },
        "groupNumber": {
          "type": "string",
          "title": "群号"
        },
        "ukey": {
          "type": "string",
          "title": "群唯一标识"
        },
        "name": {
          "type": "string",
          "title": "群名称"
        },
        "server": {
          "type": "string",
          "title": "服务器"
        },
        "avatar": {
          "type": "string",
          "title": "群头像"
        }
      },
      "title": "组织信息摘要"
    },
...
    "v1ListLeagueMembersResponse": {
      "type": "object",
      "properties": {
        "members": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1LeagueMemberSummary"
          },
          "title": "组织成员列表"
        }
      },
      "title": "获取组织成员列表响应"
    },
    "v1ListLeaguesResponse": {
      "type": "object",
      "properties": {
        "leagues": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1LeagueSummary"
          },
          "title": "组织列表"
        }
      },
      "title": "获取组织列表响应"
    },
...
    "v1SetUserNicknameResponse": {
      "type": "object",
      "title": "修改用户在组织中的昵称响应"
    },
    "v1SetUserRoleResponse": {
      "type": "object",
      "title": "设定用户在组织中的角色响应"
    },
...
    "v1UpdateLeagueExpiredAtResponse": {
      "type": "object",
      "title": "更新组织过期时间响应"
    },
    "v1UpdateLeagueResponse": {
      "type": "object",
      "title": "更新组织响应"
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Input a \"Bearer your-jwt-token\" to Value",
      "name": "Authorization",
      "in": "header"
    },
    "LeagueAuth": {
      "type": "apiKey",
      "description": "Input league key to Value Input league key to Value",
      "name": "LeagueKey",
      "in": "header"
    }
  }
}